Use off_t for sizes

+++ src/axel.h
@@ -129,7 +129,7 @@ typedef struct {
 	char filename[MAX_STRING];
 	double start_time;
 	int next_state, finish_time;
-	size_t bytes_done, start_byte, size;
+	off_t bytes_done, start_byte, size;
 	int bytes_per_second;
 	struct timespec delay_time;
 	int outfd;
